Include vulkan_headers.hpp rather than context.hpp everywhere.
