+++ application/application_wsi_events.hpp
@@ -23,7 +23,7 @@
 #pragma once
 
 #include "event.hpp"
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 
 namespace Vulkan
 {
