+++ application/platforms/application_android.cpp
@@ -25,7 +25,7 @@
 #include "application.hpp"
 #include "application_events.hpp"
 #include "application_wsi.hpp"
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include <jni.h>
 #include <android/sensor.h>
