+++ application/platforms/application_glfw.cpp
@@ -23,7 +23,7 @@
 #include "application.hpp"
 #include "application_wsi.hpp"
 #include "application_events.hpp"
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include "GLFW/glfw3.h"
 #ifdef HAVE_LINUX_INPUT
 #include "input_linux.hpp"
