+++ application/platforms/application_headless.cpp
@@ -23,7 +23,7 @@
 #include "application.hpp"
 #include "application_events.hpp"
 #include "application_wsi.hpp"
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include <thread>
 #include <mutex>
 #include <condition_variable>
