+++ application/platforms/application_libretro_utils.hpp
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include "libretro.h"
 #include "libretro_vulkan.h"
 #include "wsi.hpp"
