+++ tests/imported_host.cpp
@@ -20,7 +20,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include "device.hpp"
 #include "global_managers.hpp"
 #include <string.h>
