+++ vulkan/buffer_pool.hpp
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include "intrusive.hpp"
 #include <vector>
