+++ vulkan/device_fossilize.cpp
@@ -22,7 +22,6 @@
 
 #include "device.hpp"
 #include "timer.hpp"
-#include <stdexcept>
 
 using namespace std;
 
@@ -30,86 +29,37 @@ namespace Vulkan
 {
 void Device::register_sampler(VkSampler sampler, Fossilize::Hash hash, const VkSamplerCreateInfo &info)
 {
-	try
-	{
-		state_recorder.record_sampler(sampler, info, hash);
-	}
-	catch (const exception &e)
-	{
-		LOGE("Fossilize exception: %s\n", e.what());
-	}
+	state_recorder.record_sampler(sampler, info, hash);
 }
 
 void Device::register_descriptor_set_layout(VkDescriptorSetLayout layout, Fossilize::Hash hash, const VkDescriptorSetLayoutCreateInfo &info)
 {
-	try
-	{
-		state_recorder.record_descriptor_set_layout(layout, info, hash);
-	}
-	catch (const exception &e)
-	{
-		LOGE("Fossilize exception: %s\n", e.what());
-	}
+	state_recorder.record_descriptor_set_layout(layout, info, hash);
 }
 
 void Device::register_pipeline_layout(VkPipelineLayout layout, Fossilize::Hash hash, const VkPipelineLayoutCreateInfo &info)
 {
-	try
-	{
-		state_recorder.record_pipeline_layout(layout, info, hash);
-	}
-	catch (const exception &e)
-	{
-		LOGE("Fossilize exception: %s\n", e.what());
-	}
+	state_recorder.record_pipeline_layout(layout, info, hash);
 }
 
 void Device::register_shader_module(VkShaderModule module, Fossilize::Hash hash, const VkShaderModuleCreateInfo &info)
 {
-	try
-	{
-		state_recorder.record_shader_module(module, info, hash);
-	}
-	catch (const exception &e)
-	{
-		LOGE("Fossilize exception: %s\n", e.what());
-	}
+	state_recorder.record_shader_module(module, info, hash);
 }
 
 void Device::register_compute_pipeline(Fossilize::Hash hash, const VkComputePipelineCreateInfo &info)
 {
-	try
-	{
-		state_recorder.record_compute_pipeline(VK_NULL_HANDLE, info, nullptr, 0, hash);
-	}
-	catch (const exception &e)
-	{
-		LOGE("Fossilize exception: %s\n", e.what());
-	}
+	state_recorder.record_compute_pipeline(VK_NULL_HANDLE, info, nullptr, 0, hash);
 }
 
 void Device::register_graphics_pipeline(Fossilize::Hash hash, const VkGraphicsPipelineCreateInfo &info)
 {
-	try
-	{
-		state_recorder.record_graphics_pipeline(VK_NULL_HANDLE, info, nullptr, 0, hash);
-	}
-	catch (const exception &e)
-	{
-		LOGE("Fossilize exception: %s\n", e.what());
-	}
+	state_recorder.record_graphics_pipeline(VK_NULL_HANDLE, info, nullptr, 0, hash);
 }
 
 void Device::register_render_pass(VkRenderPass render_pass, Fossilize::Hash hash, const VkRenderPassCreateInfo &info)
 {
-	try
-	{
-		state_recorder.record_render_pass(render_pass, info, hash);
-	}
-	catch (const exception &e)
-	{
-		LOGE("Fossilize exception: %s\n", e.what());
-	}
+	state_recorder.record_render_pass(render_pass, info, hash);
 }
 
 bool Device::enqueue_create_shader_module(Fossilize::Hash hash, const VkShaderModuleCreateInfo *create_info, VkShaderModule *module)
@@ -118,7 +68,6 @@ bool Device::enqueue_create_shader_module(Fossilize::Hash hash, const VkShaderMo
 	*module = ret->get_module();
 	replayer_state.shader_map[*module] = ret;
 	return true;
-
 }
 
 void Device::notify_replayed_resources_for_type()
@@ -276,20 +225,13 @@ void Device::init_pipeline_state()
 		return;
 	}
 
-	try
-	{
-		LOGI("Replaying cached state.\n");
-		Fossilize::StateReplayer replayer;
-		auto start = Util::get_current_time_nsecs();
-		replayer.parse(*this, nullptr, static_cast<const char *>(mapped), file->get_size());
-		auto end = Util::get_current_time_nsecs();
-		LOGI("Completed replaying cached state in %.3f ms.\n", (end - start) * 1e-6);
-		replayer_state = {};
-	}
-	catch (const exception &e)
-	{
-		LOGE("Exception caught while parsing pipeline state: %s.\n", e.what());
-	}
+	LOGI("Replaying cached state.\n");
+	Fossilize::StateReplayer replayer;
+	auto start = Util::get_current_time_nsecs();
+	replayer.parse(*this, nullptr, static_cast<const char *>(mapped), file->get_size());
+	auto end = Util::get_current_time_nsecs();
+	LOGI("Completed replaying cached state in %.3f ms.\n", (end - start) * 1e-6);
+	replayer_state = {};
 }
 
 void Device::flush_pipeline_state()
