+++ vulkan/fence.hpp
@@ -23,7 +23,7 @@
 #pragma once
 
 #include "vulkan_common.hpp"
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include "object_pool.hpp"
 
 namespace Vulkan
