+++ vulkan/fence_manager.hpp
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include <vector>
 
 namespace Vulkan
