+++ vulkan/image.hpp
@@ -26,7 +26,7 @@
 #include "format.hpp"
 #include "vulkan_common.hpp"
 #include "memory_allocator.hpp"
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include <algorithm>
 
 namespace Vulkan
