+++ vulkan/pipeline_event.hpp
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "context.hpp"
+#include "vulkan_headers.hpp"
 #include "vulkan_common.hpp"
 #include "cookie.hpp"
 #include "object_pool.hpp"
